+++ changelog.md
@@ -29,6 +29,7 @@ Codion Change Log
 - EntityComboBoxModel.Builder added, related refactoring.
 - EntityComboBoxModel.ForeignKeyFilter.builder() added.
 - EntityComboBoxModel.ForeignKeyFilter.builder() now sets includeNull() to true if the foreign key is nullable.
+- EntityComboBoxModel.foreignKeyFilter() renamed filter().
 ### is.codion.swing.framework.ui
 - EntityTablePanel.Config.cellRenderer() and cellEditor() added.
 - EntityComboBox.model() added for consistency.